backend/requirements.txt
Flask
Werkzeug
Flask-Cors
Flask-SocketIO
SQLAlchemy
psycopg2-binary
python-dotenv
python-docx
docxtpl
Pillow
